fix a little bug

+++ fairseq/sequence_generator.py
@@ -138,7 +138,7 @@ class SequenceGenerator(object):
 
             # compute the encoder output for each beam
             encoder_out = model.encoder(**encoder_input)
-            new_order = torch.arange(bsz).view(-1, 1).repeat(1, beam_size).view(-1)
+            new_order = torch.arange(bsz, dtype=torch.int64).view(-1, 1).repeat(1, beam_size).view(-1)
             new_order = new_order.to(src_tokens.device)
             encoder_out = model.encoder.reorder_encoder_out(encoder_out, new_order)
             encoder_outs.append(encoder_out)
